Add cast to fix warning with gcc 4.1.
src/webkit/port/platform/graphics/svg/SVGPaintServerGradientSkia.cpp: In constructor 'WebCore::SkiaGradientBuilder::SkiaGradientBuilder(const WTF::Vector<std::pair<float, WebCore::Color>, 0u>&, float)':

src/webkit/port/platform/graphics/svg/SVGPaintServerGradientSkia.cpp:82: warning: passing 'float' for argument 1 to 'SkColor SkColorSetARGB(U8CPU, U8CPU, U8CPU, U8CPU)'

Review URL: http://codereview.chromium.org/429

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@1707 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/port/platform/graphics/svg/SVGPaintServerGradientSkia.cpp
@@ -76,7 +76,7 @@ public:
             float offset = stops[i].first;
             Color color = stops[i].second;
 
-            colors_[cur]=SkColorSetARGB(opacity*color.alpha(),
+            colors_[cur]=SkColorSetARGB(static_cast<int>(opacity*color.alpha()),
               color.red(),
               color.green(),
               color.blue());
